Remove timeout
Timeout is useful for dealing with external resources
but in this case the line parsing is only CPU bound,
so we can default to infinity.

If the user wants to limit the time spent parsing,
then it is probably best for them to wrap it in a task
and limit it for the whole document.