Don't wrap labeled wikilinks in p tags
The line scanner was recognizing the | in a wikilink label as a table
column separator. This eventually still got turned into a link, but by
then, it was already wrapped in a Para block.

To fix this, eat anything wrapped in double square brackets before
looking for the table regex.