I112 bad para in wikilinks (#114)
* Don't wrap labeled wikilinks in p tags (#113)

The line scanner was recognizing the | in a wikilink label as a table
column separator. This eventually still got turned into a link, but by
then, it was already wrapped in a Para block.

To fix this, eat anything wrapped in double square brackets before
looking for the table regex.

Co-authored-by: Ben Olive <sionide21@gmail.com>

* Added RELEASE NOTES for 1.4.28

* modernised wikilinks tests

* testing all: true implemented, bug for <sup> fixed

* fixing doctest for 1.11

* Prepare rel 1.4.28

Co-authored-by: Ben Olive <sionide21@gmail.com>