Avoid deprecated negative slice in Elixir v1.16-dev
There is still a long time for Elixir v1.16
but since I run on the edge, I thought I
would send a PR.

Another solution for the problem is to
use `String.slice(string, ignore..-1//1)`
but that would require Elixir v1.12. The
version in this PR supports earlier
versions too.